Spring2 market lesson 7

+++ api/src/main/java/ru/senkin/marker/api/dto/OrderDto.java
@@ -1,11 +1,12 @@
 package ru.senkin.marker.api.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class OrderDto {
     private Long orderId;
     private List<OrderItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public Long getOrderId() {
         return orderId;
@@ -15,7 +16,7 @@ public class OrderDto {
         return items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
@@ -27,14 +28,14 @@ public class OrderDto {
         this.items = items;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
     public OrderDto() {
     }
 
-    public OrderDto(Long orderId, List<OrderItemDto> items, int totalPrice) {
+    public OrderDto(Long orderId, List<OrderItemDto> items, BigDecimal totalPrice) {
         this.orderId = orderId;
         this.items = items;
         this.totalPrice = totalPrice;
